+++ Pandora/Pattern/Junction/Transformer.hs
@@ -11,5 +11,8 @@ class Composition t => Transformer t where
 	lay :: Covariant u => u ~> Schema t u
 	wrap :: Pointable u => t ~> Schema t u
 
-infixr 1 :>
-type (:>) t u a = Transformer t => Schema t u a
+-- infixr 1 :>
+-- type (:>) t u a = Transformer t => Schema t u a
+
+infixr 3 :>
+newtype (:>) t u a = T { trans :: Transformer t => Schema t u a }