+++ CHANGELOG.md
@@ -118,3 +118,4 @@
 * Define `Cartesian` type class
 * Rename `?` to `%` to use `?` as boolean multi-if
 * Replace `ifelse` method from `Setoid` module
+* Convert `:>` to a newtype